Get_twitter_v1/requirements.txt
flask
python-dotenv
mysql-connector-python
httpx